third_party/blink/web_tests/platform/generic/fast/xpath/4XPath/Core/test_node_test-expected.txt

PASS nodeInResult(ROOT, result) is true
PASS nodeInResult(ROOT, result) is false
PASS nodeInResult(CHILD1, result) is false
PASS nodeInResult(CHILD3, result) is true
PASS nodeInResult(ROOT, result) is false
PASS nodeInResult(CHILD1, result) is false
PASS nodeInResult(CHILD3, result) is true
PASS nodeInResult(ROOT, result) is true
PASS nodeInResult(TEXT1, result) is true
PASS nodeInResult(ROOT, result) is false
PASS nodeInResult(TEXT1, result) is true
PASS nodeInResult(ROOT, result) is false
PASS nodeInResult(COMMENT, result) is true
PASS nodeInResult(ROOT, result) is false
PASS nodeInResult(PI, result) is true
PASS nodeInResult(PI, result) is true
PASS nodeInResult(PI2, result) is false
PASS successfullyParsed is true

TEST COMPLETE
